predictor/pyproject.toml
[project]
name = "predictor"
version = "1.1.0"
description = "Vertex AI | Custom Model Serving | AutoGluon"
requires-python = ">=3.12"
dependencies = [
    "autogluon-tabular[all]>=1.3.1",
    "autogluon-multimodal==1.3.1",
    "google-cloud-storage>=3.1.0",
    "litestar[structlog]>=2.16.0",
    "pandas>=2.2.3",
    "uvicorn>=0.34.3",
    "structlog-gcp>=0.5.0",
]

[project.optional-dependencies]
cpu = ["torch==2.6.0", "torchvision==0.21.0"]
gpu = ["torch==2.6.0", "torchvision==0.21.0"]

[dependency-groups]
dev = [
    "bandit>=1.8.0",
    "mypy>=1.13.0",
    "pytest>=8.3.5",
    "pytest-cov>=6.2.1",
    "pytest-mock>=3.14.0",
    "ruff>=0.12.1",
]

[tool.ruff]
line-length = 80

[tool.mypy]
ignore_missing_imports = true
disable_error_code = ["import-untyped", "attr-defined"]

[tool.bandit]
exclude_dirs = ["tests"]

[tool.uv]
conflicts = [[{ extra = "cpu" }, { extra = "gpu" }]]

[tool.uv.sources]
torch = [
    { index = "pytorch-cpu", extra = "cpu" },
    { index = "pytorch-gpu", extra = "gpu" },
]
torchvision = [
    { index = "pytorch-cpu", extra = "cpu" },
    { index = "pytorch-gpu", extra = "gpu" },
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[[tool.uv.index]]
name = "pytorch-gpu"
url = "https://download.pytorch.org/whl/cu126"
explicit = true


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
